--- src/finder/libs/file.py
@@ -1,5 +1,0 @@
-class File():
-    def __init__(self, path):
-        self.name = path.split('\\')[-1]
-        self.path = path
-        self.score = 0
