--- src/finder/libs/folder.py
@@ -1,5 +1,0 @@
-class Folder():
-    def __init__(self, path):
-        self.name = path.split('\\')[-1]
-        self.path = path
-        self.score = 0
