+++ src/finder/scores.py
@@ -1,51 +1,51 @@
 import re 
 
-def updateScores(basic_rx, sub_rx, folders):
+def updateScores(basic_rx, sub_rx, index):
     temp = []
-    for folder in folders:
-        folder.score = getScore(basic_rx, sub_rx, folder.name.lower())
-        #folder.score += getScore(basic_rx, sub_rx, folder.name.lower()) removed for better results on all match
+    for item in index:
+        item.score = getScore(basic_rx, sub_rx, item.name.lower())
 
-        if folder.score > 0:
-            temp.append(folder)
+        if item.score > 0:
+            temp.append(item)
 
     return temp
 
-def advancedScoreUpdate(q, folders):
+def advancedScoreUpdate(q, index):
     j = 0
-    for f in folders:
+    for f in index:
         w = ""
         for i in re.split(' |\.|-|_', f.name.lower()):
             if len(i) > 0:
                 w += i[0]
         for c in q:
             if c in w:
-                folders[j].score = folders[j].score + 10
-        folders[j].score = folders[j].score - len(re.split(' |\.|-|_', f.name.lower()))
-    
+                index[j].score = index[j].score + 10
+
+        index[j].score = index[j].score - len(re.split(' |\.|-|_', f.name.lower()))
         j += 1   
 
-    return folders
+    return index
 
-def spaceScores(q, r, folders):
+def spaceScores(q, r, index):
     if ' ' not in q:
-        return folders
+        return index
     
     temp = []
-    for folder in folders:
-        string = folder.path.split('\\')[:-1]
+    for item in index:
+        string = item.path.split('\\')[:-1]
         string = ' '.join(string)
 
         match = r.search(string.lower())
 
         if not match == None:
-            folder.score += 10
-            temp.append(folder)
+            item.score += 10
+            temp.append(item)
 
     return temp
 
 
 def getScore(basic_rx, sub_rx, string):
+
     match_basic = basic_rx.search(string)
     match_sub = sub_rx.search(string)
